+++ R/zzz.R
@@ -1,5 +1,7 @@
-#' @importFrom glue glue 
+#' @importFrom bibtex read.bib write.bib
+#' @importFrom cli cli_alert_info
+#' @importFrom fs path_ext
+#' @importFrom glue glue
 #' @importFrom rcrossref cr_cn
-#' @importFrom RefManageR ReadBib
 #' @importFrom stringr str_locate str_detect
 NULL