+++ NEWS.md
@@ -1,6 +1,9 @@
-# inSilecoRef (devel)
+# inSilecoRef DEVEL
 
+* Re-write `newDOI()` that is now named `add_doi_reference()`.
 * Add new functions to quickly generate commonly used URL for R packages: `url_cran()`, `url_bioc()` and `url_gh()`.
+
+
 
 # inSilecoRef 0.0.1